# Do not allocate individual nodes and ways. Try to optimize tags.