@paulmach 2020-10-21
Thank you, I got similar speedup using 1.15.2

```
benchmark             old ns/op     new ns/op     delta
BenchmarkDecode-4     799414260     737218287     -7.78%
BenchmarkLondon-4     834297009     740366127     -11.26%

benchmark             old allocs     new allocs     delta
BenchmarkDecode-4     7732796        4545134        -41.22%
BenchmarkLondon-4     7732796        4545134        -41.22%

benchmark             old bytes      new bytes      delta
BenchmarkDecode-4     1306693687     1305443549     -0.10%
BenchmarkLondon-4     1306693457     1305443541     -0.10%
```